requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
pandas
numpy
scikit-learn
joblib
catboost
